+++ app/src/main/java/com/example/rss_aggregator_2022/features/presentation/rssmanagement/RssManagementFragment.kt
@@ -1,21 +1,21 @@
-package com.example.rss_aggregator_2022.features.rssfeed
+package com.example.rss_aggregator_2022.features.presentation.rssmanagement
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.rss_aggregator_2022.databinding.FragmentRssFeedBinding
+import com.example.rss_aggregator_2022.databinding.FragmentRssManagementBinding
 
-class RssFeedFragment : Fragment() {
-    var binding: FragmentRssFeedBinding? = null
+class RssManagementFragment : Fragment() {
+    var binding: FragmentRssManagementBinding? = null
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentRssFeedBinding.inflate(inflater)
+        binding = FragmentRssManagementBinding.inflate(inflater)
         return binding?.root
     }
 }
