+++ app/src/main/java/com/example/rss_aggregator_2022/features/presentation/rssmanagement/UserFormFragment.kt
@@ -1,4 +1,4 @@
-package com.example.rss_aggregator_2022.features.rssmanagement
+package com.example.rss_aggregator_2022.features.presentation.rssmanagement
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,16 +6,18 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.rss_aggregator_2022.databinding.FragmentRssManagementBinding
+import com.example.rss_aggregator_2022.databinding.UserFormViewBinding
 
-class RssManagementFragment : Fragment() {
-    var binding: FragmentRssManagementBinding? = null
+class UserFormFragment:Fragment() {
+    private var binding: UserFormViewBinding? = null
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentRssManagementBinding.inflate(inflater)
+        binding = UserFormViewBinding.inflate(inflater)
         return binding?.root
     }
+
 }
