+++ app/src/main/res/layout/fragment_rss_management.xml
@@ -12,5 +12,5 @@
         app:layout_constraintTop_toTopOf="parent"
         android:elevation="@dimen/elevation"
         app:title="@string/manager_name"
-        app:menu="@menu/management_menu"/>/>
+        app:menu="@menu/management_menu"/>
 </androidx.constraintlayout.widget.ConstraintLayout>
