+++ app/src/main/res/menu/management_menu.xml
@@ -2,7 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
     <item
-        android:id="@+id/to_rss_feed_item"
+        android:id="@+id/to_user_form"
         android:icon="@drawable/ic_plus"
         android:title="@string/Add_option"
         app:showAsAction="ifRoom" />
