+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/feed"
-        android:name="com.example.rss_aggregator_2022.features.rssfeed.RssFeedFragment"
+        android:name="com.example.rss_aggregator_2022.features.presentation.rssfeed.RssFeedFragment"
         android:label="@string/feed_name"
         tools:layout="@layout/fragment_rss_feed">
 
@@ -15,16 +15,29 @@
     <fragment
         android:id="@+id/management"
         android:label="@string/manager_name"
-        android:name="com.example.rss_aggregator_2022.features.rssmanagement.RssManagementFragment"
+        android:name="com.example.rss_aggregator_2022.features.presentation.rssmanagement.RssManagementFragment"
         tools:layout="@layout/fragment_rss_management">
+
+        <action android:id="@+id/action_management_to_user_form"
+            app:destination="@+id/user_form"/>
 
     </fragment>
     <fragment
         android:id="@+id/profile"
-        android:name="com.example.rss_aggregator_2022.features.profile.ProfileFragment"
+        android:name="com.example.rss_aggregator_2022.features.presentation.profile.ProfileFragment"
         android:label="@string/profile_name"
         tools:layout="@layout/fragment_profile">
 
     </fragment>
+
+    <dialog
+        android:id="@+id/user_form"
+        android:name="com.example.rss_aggregator_2022.features.presentation.rssmanagement.UserFormFragment">
+
+        <action
+            android:id="@+id/action_user_form_to_management"
+            app:destination="@+id/management"/>
+
+    </dialog>
 
 </navigation>
